requirements.txt
Django==2.2.*
django-bootstrap-pagination==1.7.0
django-bootstrap3==14.2.0
django-bootstrap3-datetimepicker-2==2.8.2
django-common-helpers==0.9.2
django-cron==0.5.1
django-dual-authentication==1.2.1
django-registration-redux==2.9
django-recurrence==1.10.3
